+++ test/index.test.js
@@ -1,0 +1,10 @@
+// @flow
+
+import "../src/index"
+
+declare var describe: Function
+declare var it: Function
+
+describe("stub for checking coverage", () => {
+	it("shouldn't do anything", () => {})
+})