+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 
-
-
+- added [useLeafletIsMoving](README.md#useleafletismoving) & [useLeafletIsZooming](README.md#useleafletiszooming) hooks.